physics bug fixes

+++ include/rdge/physics/collision.hpp
@@ -30,7 +30,7 @@ static constexpr float LINEAR_SLOP = 0.005f; //!< Collision tolerance
 //! \brief Container for collision resolution details
 struct collision_manifold
 {
-    uint32 count = 0;               //!< Number of collision points
+    size_t count = 0;               //!< Number of collision points
     float depths[2] = { 0.f, 0.f }; //!< Penetration depths
     math::vec2 contacts[2];         //!< Contact points
     math::vec2 normal;              //!< Vector of resolution, or collision normal
